requirements.txt
py-midi
numpy
tensorflow